[WCNCR00262566] wlan: refacotr pointer casting for calculation (2/.)
[Description]
refacotr pointer casting for calculation

MTK-Commit-Id: ad94f99039a52a6ee93e4b7529141abe8d3a5cdf

CR-Id: WCNCR00262566
Feature: wlan/os
Reviewed-on: https://gerrit.mediatek.inc/c/neptune/wlan_driver/gen4m/+/6096477
CheckCourse: srv_pis <srv_pis@mediatek.com>
CheckCRInfo: srv_pis <srv_pis@mediatek.com>
CheckForbiddenKeyword: srv_pis <srv_pis@mediatek.com>
CheckPatch: srv_pis <srv_pis@mediatek.com>
Tested-by: yuchi.wang <yuchi.wang@mediatek.com>
Build: srv_neptune_adm <srv_neptune_adm@mediatek.com>
Build: srv_preflight_a001 <srv_preflight_a001@mediatek.com>
Test: srv_mspautosanity

+++ mgmt/bss.c
@@ -1212,14 +1212,13 @@ struct MSDU_INFO* bssComposeBeaconContent(IN struct ADAPTER *prAdapter,
 
 	/* 4 <2> Compose header */
 	bssComposeBeaconProbeRespFrameHeaderAndFF((uint8_t *)
-						  ((unsigned
-						    long)(prMsduInfo->prPacket)
-						   + MAC_TX_RESERVED_FIELD),
-						  NULL,
-						  prBssInfo->aucOwnMacAddr,
-						  prBssInfo->aucBSSID,
-						  prBssInfo->u2BeaconInterval,
-						  prBssInfo->u2CapInfo);
+		((uintptr_t)(prMsduInfo->prPacket)
+			+ MAC_TX_RESERVED_FIELD),
+			NULL,
+			prBssInfo->aucOwnMacAddr,
+			prBssInfo->aucBSSID,
+			prBssInfo->u2BeaconInterval,
+			prBssInfo->u2CapInfo);
 
 	prMsduInfo->u2FrameLength = (WLAN_MAC_MGMT_HEADER_LEN +
 				     (TIMESTAMP_FIELD_LEN +
